Lift the AST unparser into the main crate
Move `fuzz/mutator/src/unparse.rs` to `aver::ast::unparse` so the
compiler's proof-explain renderer and the fuzz mutator share one
AST -> Aver-source unparser (a single round-trip target). The fuzz
module becomes a thin re-export, leaving every `crate::unparse::…`
call site unchanged. Adds a public `write_expr_public` entry so a
single expression can be rendered on its own.
